Refactored BinarySerach and Hashing
