+++ src/BinarySearch/NextAlphabeticalElement.java
@@ -1,15 +1,17 @@
-public class CeilOfAnElement
+package BinarySearch;
+
+public class NextAlphabeticalElement
 {
-    public int ceilOfAnElement(int[] a,int element)
+    public char nextAlphabeticalElement(char[] a,char ch)
     {
         int l=0;
         int r=a.length-1;
-        int result=-1;
+        char result='0';
 
         while(l<=r)
         {
-            int mid = l+(r-l)/2;
-            if(a[mid]>element)
+            int mid=l+(r-l)/2;
+            if(a[mid]>ch)
             {
                 result=a[mid];
                 r=mid-1;
@@ -17,7 +19,6 @@ public class CeilOfAnElement
             else {
                 l=mid+1;
             }
-
         }
         return result;
     }
